ready for deployment

+++ P2/flask_api/resources/artist.py
@@ -1,5 +1,6 @@
 from flask_restful import Resource, reqparse
 from db import db
+from lock import lock
 from models.account import auth
 from models.artist import ArtistModel
 from models.event import EventModel
@@ -16,7 +17,7 @@ class ArtistEventsList(Resource):
         """
         artist = ArtistModel.find_by_id(id)
         if not artist:
-            return {'message': "Artist with id {} not found".format(id)}, 404
+            return {'message': "Artist with ['id': {}] not found".format(id)}, 404
         events = [a.json()['event'] for a in db.session.query(EventModel).all() if a.artist_in_event(artist.name)]
         return {'events': events}, 200
 
@@ -44,7 +45,7 @@ class Artist(Resource):
         artist = ArtistModel.find_by_id(id)
         if artist:
             return artist.json(), 200
-        return {'message': "Artist with id {} not found".format(id)}, 404
+        return {'message': "Artist with ['id': {}] not found".format(id)}, 404
 
     @auth.login_required(role='admin')
     def post(self, id=None):
@@ -53,14 +54,17 @@ class Artist(Resource):
         used already (this was a requirement on one of the first sessions).
 
         :param id: artist ID [OPTIONAL]
-        :return: artist info in a JSON structure (200) | error ID already taken (409)
+        :return: artist info in a JSON structure (201) | error ID already taken (409)
         """
-        if ArtistModel.find_by_id(id):
-            return {'message': "An artist with id {} already exists".format(id)}, 409
         data = self.__parse_request__()
-        new_artist = ArtistModel(data.get('name'), data.get('country'), data.get('genre'), id)
-        new_artist.save_to_db()
-        return new_artist.json(), 200
+        with lock.lock:
+            art = ArtistModel.find_by_name(data.get('name'))
+            if art:
+                return {'message': "An artist with ['name': {}] already exists".format(art.name)}, 409
+            new_artist = ArtistModel(" ".join(w.capitalize() for w in data.get('name').split(" ")),
+                                     data.get('country'), data.get('genre'), id)
+            new_artist.save_to_db()
+            return new_artist.json(), 201
 
     @auth.login_required(role='admin')
     def delete(self, id):
@@ -70,11 +74,12 @@ class Artist(Resource):
         :param id: artist ID
         :return: message ok (200) | error not found (404)
         """
-        to_delete = ArtistModel.find_by_id(id)
-        if not to_delete:
-            return {'message': "There is no artist with id {}, therefore it cannot be deleted".format(id)}, 404
-        to_delete.delete_from_db()
-        return {'message': "Artist with id {} has benn successfully been deleted".format(id)}, 200
+        with lock.lock:
+            to_delete = ArtistModel.find_by_id(id)
+            if not to_delete:
+                return {'message': "There is no artist with ['id': {}], therefore it cannot be deleted".format(id)}, 404
+            to_delete.delete_from_db()
+            return {'message': "Artist with ['name': {}, 'id': {}] has successfully been deleted".format(to_delete.name, id)}, 200
 
     @auth.login_required(role='admin')
     def put(self, id):
@@ -83,15 +88,22 @@ class Artist(Resource):
         given data.
 
         :param id: artist ID
-        :return: created or modified artist info in a JSON structure (200)
+        :return: created or modified artist info in a JSON structure (200) | bad request (400) | artist not found (404)
         """
         data = self.__parse_request__()
-        existing = ArtistModel.find_by_id(id)
+        existing = ArtistModel.find_by_name(data.get('name'))
         if existing:
+            return {"message": "Artist name modification not allowed, artist with ['name': {}] "
+                               "already exists".format(existing.name)}, 400
+        with lock.lock:
+            existing = ArtistModel.find_by_id(id)
+            if not existing:
+                return {"message": "Artist with ['id': {}] not found".format(id)}, 404
             existing.delete_from_db()
-        new_artist = ArtistModel(data.get('name'), data.get('country'), data.get('genre'), id)
-        new_artist.save_to_db()
-        return new_artist.json(), 200
+            new_artist = ArtistModel(" ".join(w.capitalize() for w in data.get('name').split(" ")),
+                                     data.get('country'), data.get('genre'), id)
+            new_artist.save_to_db()
+            return new_artist.json(), 200
 
     def __parse_request__(self):
         """
@@ -100,7 +112,7 @@ class Artist(Resource):
         :return: parsed data
         """
         parser = reqparse.RequestParser()
-        parser.add_argument('name', type=str, required=True, help="This field cannot be left blank")
-        parser.add_argument('country', type=str)
-        parser.add_argument('genre', type=str)
+        parser.add_argument('name', type=str, required=True, help="Operation not valid: 'name' not provided")
+        parser.add_argument('country', type=str, required=True, help="Operation not valid: 'country' not provided")
+        parser.add_argument('genre', type=str, required=True, help="Operation not valid: 'genre' not provided")
         return parser.parse_args()
